% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsPanelApp.R
\name{runNGSPanelApp}
\alias{runNGSPanelApp}
\title{Run the NGS Panel Shiny app}
\usage{
runNGSPanelApp()
}
\description{
This function runs the main shiny app. PRIOR to running this
function, the user will have needed to download the COSMIC tsv
file. That file must currently be called "cosmic.tsv" and live
in the current working directory.
}
\details{
#@param cosmic_tsv The path to the manually-downloaded cosmic tsv
#    file.
}
